+++ src/openapi_mutator/add_request.rs
@@ -116,3 +116,40 @@ fn field_names(api: &OpenAPI, request_body: &RequestBody) -> Option<Vec<String>>
         _ => None,
     }
 }
+
+#[cfg(test)]
+mod test {
+    use libafl::mutators::{MutationResult, Mutator};
+
+    use super::AddRequestMutator;
+    use crate::{
+        input::{Method, OpenApiInput},
+        state::tests::TestOpenApiFuzzerState,
+    };
+
+    /// Tests whether the mutator adds a valid request (including parameters, if required for the chosen request).
+    #[test]
+    fn add_correct_request() -> anyhow::Result<()> {
+        for _ in 0..100 {
+            let mut state = TestOpenApiFuzzerState::new();
+            let mut input = OpenApiInput(vec![]);
+            let mut mutator = AddRequestMutator;
+
+            let result = mutator.mutate(&mut state, &mut input)?;
+            assert_eq!(result, MutationResult::Mutated);
+            assert_eq!(input.0.len(), 1);
+            assert!(TestOpenApiFuzzerState::PATHS.contains(&input.0[0].path.as_str()));
+            assert!(
+                input.0[0].method == Method::Get
+                    || (input.0[0].path == "/simple" && input.0[0].method == Method::Delete)
+            );
+            if input.0[0].path == "/with-query-parameter"
+                || input.0[0].path == "/with-path-parameter/{id}"
+            {
+                assert!(input.0[0].contains_parameter("id"));
+            }
+        }
+
+        Ok(())
+    }
+}