+++ src/openapi_mutator/break_link.rs
@@ -64,3 +64,35 @@ where
         Ok(())
     }
 }
+
+#[cfg(test)]
+mod test {
+    use libafl::mutators::{MutationResult, Mutator};
+
+    use super::BreakLinkMutator;
+    use crate::{
+        input::parameter::ParameterKind, openapi_mutator::test_helpers::linked_requests,
+        state::tests::TestOpenApiFuzzerState,
+    };
+
+    /// Tests whether the mutator correctly breaks a reference parameter.
+    #[test]
+    fn break_link() -> anyhow::Result<()> {
+        for _ in 0..100 {
+            let mut state = TestOpenApiFuzzerState::new();
+            let mut input = linked_requests();
+            let mut mutator = BreakLinkMutator;
+
+            let result = mutator.mutate(&mut state, &mut input)?;
+            assert_eq!(result, MutationResult::Mutated);
+            assert!(
+                input.0[1]
+                    .get_mut_parameter("id", ParameterKind::Query)
+                    .expect("Could not find parameter after request removal")
+                    .bytes()
+                    .is_some()
+            );
+        }
+        Ok(())
+    }
+}