+++ src/openapi_mutator/different_path.rs
@@ -83,3 +83,38 @@ where
         Ok(())
     }
 }
+
+#[cfg(test)]
+mod test {
+    use libafl::mutators::{MutationResult, Mutator};
+
+    use super::DifferentPathMutator;
+    use crate::{
+        input::Method, openapi_mutator::test_helpers::simple_request,
+        state::tests::TestOpenApiFuzzerState,
+    };
+
+    /// Tests whether the mutator correctly changes the path of a request.
+    #[test]
+    fn mutate_path() -> anyhow::Result<()> {
+        for _ in 0..100 {
+            let mut state = TestOpenApiFuzzerState::new();
+
+            let mut input = simple_request();
+            input.0[0].path = "/with-query-parameter".to_string();
+
+            let mut mutator = DifferentPathMutator;
+            let result = mutator.mutate(&mut state, &mut input)?;
+
+            let new_path = input.0[0].path.as_str();
+            assert!(new_path == "/with-path-parameter/{id}" || new_path == "/simple");
+            assert!(
+                input.0[0].method == Method::Get
+                    || (new_path == "/simple" && input.0[0].method == Method::Delete)
+            );
+            assert_eq!(result, MutationResult::Mutated);
+        }
+
+        Ok(())
+    }
+}