+++ src/openapi_mutator/duplicate_request.rs
@@ -71,3 +71,74 @@ where
         Ok(())
     }
 }
+
+#[cfg(test)]
+mod test {
+    use libafl::mutators::{MutationResult, Mutator};
+
+    use super::DuplicateRequestMutator;
+    use crate::{
+        input::{OpenApiInput, parameter::ParameterKind},
+        openapi_mutator::test_helpers::{linked_requests, simple_request},
+        state::tests::TestOpenApiFuzzerState,
+    };
+
+    /// Tests whether the mutator correctly skips mutation if there's no requests.
+    #[test]
+    fn skip_when_empty() -> anyhow::Result<()> {
+        for _ in 0..100 {
+            let mut state = TestOpenApiFuzzerState::new();
+            let mut input = OpenApiInput(vec![]);
+            let mut mutator = DuplicateRequestMutator;
+
+            let result = mutator.mutate(&mut state, &mut input)?;
+            assert_eq!(result, MutationResult::Skipped);
+        }
+
+        Ok(())
+    }
+
+    /// Tests whether the mutator correctly duplicates a "simple" request (i.e. no parameters or body).
+    #[test]
+    fn duplicate_request_simple() -> anyhow::Result<()> {
+        for _ in 0..100 {
+            let mut state = TestOpenApiFuzzerState::new();
+            let mut input = simple_request();
+            let mut mutator = DuplicateRequestMutator;
+
+            let result = mutator.mutate(&mut state, &mut input)?;
+            assert_eq!(result, MutationResult::Mutated);
+            assert_eq!(input.0[0].path, input.0[1].path);
+            assert_eq!(input.0[0].method, input.0[1].method);
+            assert_eq!(
+                input.0[0].body_content_type(),
+                input.0[1].body_content_type()
+            );
+        }
+
+        Ok(())
+    }
+
+    /// Tests whether the mutator correctly duplicates a request containing a parameter reference.
+    #[test]
+    fn duplicate_request_with_reference() -> anyhow::Result<()> {
+        for _ in 0..100 {
+            let mut state = TestOpenApiFuzzerState::new();
+            let mut input = linked_requests();
+            let mut mutator = DuplicateRequestMutator;
+
+            let result = mutator.mutate(&mut state, &mut input)?;
+            assert_eq!(result, MutationResult::Mutated);
+            let new_request = &mut input.0[2];
+            if new_request.path == "/with-query-parameter" {
+                let parameter = new_request
+                    .get_mut_parameter("id", ParameterKind::Query)
+                    .expect("Parameter was not correctly duplicated");
+                assert!(parameter.is_reference());
+                assert_eq!(parameter.reference_index().copied(), Some(0));
+            }
+        }
+
+        Ok(())
+    }
+}